pyproject.toml
[project]
name = "pi2w-server"
version = "0.1.0"
description = "Content server for Inky Impression e-ink display"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "hitherdither @ git+https://github.com/hbldh/hitherdither.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
